create template pages in separate plugin
