+++ gridsome/lib/plugins/source-vue/index.js
@@ -2,7 +2,6 @@ const path = require('path')
 const fs = require('fs-extra')
 const glob = require('globby')
 const slash = require('slash')
-const { find } = require('lodash')
 const chokidar = require('chokidar')
 const { createPagePath, parseComponent } = require('./lib/utils')
 
@@ -11,13 +10,9 @@ class VueSource {
     return {}
   }
 
-  constructor (api, options) {
-    this.options = options
-    this.store = api.store
+  constructor (api) {
     this.pages = api.pages
-
     this.pagesDir = api.config.pagesDir
-    this.templatesDir = api.config.templatesDir
 
     api.transpileDependencies([path.resolve(__dirname, 'lib', 'loaders')])
     api.registerComponentParser({ test: /\.vue$/, parse: parseComponent })
@@ -29,10 +24,6 @@ class VueSource {
 
     if (fs.existsSync(this.pagesDir)) {
       api.createPages(args => this.createPages(args))
-    }
-
-    if (fs.existsSync(this.templatesDir)) {
-      api.createPages(args => this.createTemplates(args))
     }
   }
 
@@ -84,95 +75,6 @@ class VueSource {
   removePage (file) {
     const component = path.join(this.pagesDir, file)
     return this.pages.removePage({ component })
-  }
-
-  async createTemplates () {
-    const files = await glob('**/*.vue', { cwd: this.templatesDir })
-
-    for (const file of files) {
-      this.createTemplate(file)
-    }
-
-    if (process.env.NODE_ENV === 'development') {
-      const watcher = chokidar.watch('**/*.vue', {
-        ignoreInitial: true,
-        cwd: this.templatesDir
-      })
-
-      watcher.on('add', file => this.createTemplate(slash(file)))
-      watcher.on('unlink', file => this.removeTemplate(slash(file)))
-    }
-  }
-
-  createTemplate (file) {
-    const component = path.join(this.templatesDir, file)
-    const contentType = find(this.store.store.collections, ({ options }) => {
-      return options.component === component
-    })
-
-    if (!contentType) return
-
-    contentType.collection.find().forEach(node => {
-      this.createNodePage(node)
-    })
-
-    if (process.env.NODE_ENV === 'development') {
-      contentType.on('add', this.createNodePage, this)
-      contentType.on('remove', this.removeNodePage, this)
-      contentType.on('update', this.updateNodePage, this)
-    }
-  }
-
-  removeTemplate (file) {
-    const component = path.join(this.templatesDir, file)
-    const contentType = find(this.store.store.collections, ({ options }) => {
-      return options.component === component
-    })
-
-    if (!contentType) return
-
-    this.pages.removePage({ component })
-
-    if (process.env.NODE_ENV === 'development') {
-      contentType.off('add', this.createNodePage, this)
-      contentType.off('remove', this.removeNodePage, this)
-      contentType.off('update', this.updateNodePage, this)
-    }
-  }
-
-  createNodePage (node) {
-    const contentType = this.store.getContentType(node.typeName)
-    const component = path.join(this.templatesDir, `${node.typeName}.vue`)
-    const { route } = contentType.options
-
-    return this.pages.createPage({
-      queryContext: node,
-      path: node.path,
-      component,
-      route
-    })
-  }
-
-  updateNodePage (node, oldNode) {
-    const contentType = this.store.getContentType(node.typeName)
-    const component = path.join(this.templatesDir, `${node.typeName}.vue`)
-    const { route } = contentType.options
-
-    if (node.path !== oldNode.path && !route) {
-      this.removeNodePage(oldNode)
-      return this.createNodePage(node)
-    }
-
-    return this.pages.createPage({
-      queryContext: node,
-      path: node.path,
-      component,
-      route
-    })
-  }
-
-  removeNodePage (node) {
-    return this.pages.removePage({ path: node.path })
   }
 }
